requirements.txt
# =============================================================================
# TELEGRAM BOT FRAMEWORK
# =============================================================================
python-telegram-bot>=20.0

# =============================================================================
# GEMINI AI INTEGRATION
# =============================================================================
google-generativeai>=0.3.0

# =============================================================================
# GOOGLE SERVICES API INTEGRATION
# =============================================================================
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# =============================================================================
# DOCUMENT PROCESSING
# =============================================================================
PyPDF2>=3.0.0
python-docx>=1.1.0

# =============================================================================
# SLACK INTEGRATION
# =============================================================================
slack_sdk>=3.0.0

# =============================================================================
# NOTION INTEGRATION
# =============================================================================
notion-client>=2.0.0

# =============================================================================
# HTTP REQUESTS
# =============================================================================
requests>=2.31.0

# =============================================================================
# ENVIRONMENT VARIABLE MANAGEMENT
# =============================================================================
python-dotenv>=1.0.0

# =============================================================================
# IMAGE PROCESSING
# =============================================================================
Pillow>=10.0.0

# =============================================================================
# AUDIO PROCESSING
# =============================================================================
pydub>=0.25.0

# =============================================================================
# FILE PROCESSING UTILITIES
# =============================================================================
python-magic>=0.4.27

# =============================================================================
# EXTERNAL DEPENDENCIES (MANUAL INSTALLATION REQUIRED)
# =============================================================================
# ffmpeg - For audio processing
#   Ubuntu/Debian: sudo apt-get install ffmpeg
#   Windows: Download from https://ffmpeg.org/download.html
#   macOS: brew install ffmpeg
